+++ examples/triangle/src/androidMain/kotlin/AppResources.kt
@@ -25,6 +25,8 @@ import graphics.glimpse.meshes.ArrayMeshData
 import graphics.glimpse.meshes.obj.ObjMeshDataParser
 import graphics.glimpse.meshes.obj.parseArrayMeshData
 import graphics.glimpse.shaders.ShaderType
+import graphics.glimpse.textures.TextureImageSource
+import graphics.glimpse.textures.fromAsset
 import java.lang.ref.WeakReference
 import java.util.*
 
@@ -67,7 +69,25 @@ actual class AppResources(context: Context) {
     private fun getShaderAssetPath(type: ShaderType): String =
         "${type.name.toLowerCase(Locale.ENGLISH)}.glsl"
 
+    /**
+     * Returns texture image source.
+     */
+    actual fun getTextureSource(): TextureImageSource =
+        TextureImageSource.builder()
+            .fromAsset(context, TEXTURE_FILE_NAME)
+            .build()
+
+    /**
+     * Returns normal map image source.
+     */
+    actual fun getNormalMapSource(): TextureImageSource =
+        TextureImageSource.builder()
+            .fromAsset(context, NORMAL_MAP_FILE_NAME)
+            .build()
+
     companion object {
         private const val MESH_FILE_NAME = "triangle.obj"
+        private const val TEXTURE_FILE_NAME = "bricks_texture.png"
+        private const val NORMAL_MAP_FILE_NAME = "bricks_normalmap.png"
     }
 }