+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Java resource texture image source (desktop)
+
 ### Fixed
 - Publish Android artifacts for `debug` variant
+- Fix IndexOutOfBoundsException in texture builder
+- Fix loading texture image (desktop)
 
 ## [1.0.0-ALPHA1]
 ### Added